refactor(raspberry): split helpers.js into focused single-responsibility modules
The monolithic helpers.js files mixed paths, shell execution, sanitization,
formatting, and video config logic. Split into dedicated modules for better
readability and maintainability. Original helpers.js files remain as re-export
shims for backward compatibility.

- admin/paths.js — path constants and config
- admin/shell-exec.js — secure shell command execution
- admin/sanitize.js — filename/segment sanitization
- admin/formatting.js — uptime and disk info formatting
- admin/video-config.js — video config helpers and NEOPRO locking
- server/env-config.js — environment variables and file paths
