+++ raspberry/admin/formatting.js
@@ -1,0 +1,31 @@
+/**
+ * Fonctions de formatage : uptime, informations disque.
+ */
+
+function formatUptime(seconds) {
+  const days = Math.floor(seconds / 86400);
+  const hours = Math.floor((seconds % 86400) / 3600);
+  const minutes = Math.floor((seconds % 3600) / 60);
+
+  const parts = [];
+  if (days > 0) parts.push(`${days}j`);
+  if (hours > 0) parts.push(`${hours}h`);
+  parts.push(`${minutes}m`);
+
+  return parts.join(' ');
+}
+
+function parseDiskInfo(output) {
+  const parts = output.split(/\s+/);
+  return {
+    total: parts[1],
+    used: parts[2],
+    available: parts[3],
+    percent: parts[4],
+  };
+}
+
+module.exports = {
+  formatUptime,
+  parseDiskInfo,
+};
